refactor config into its own module
